backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
google-generativeai
openai
requests
python-jose[cryptography]
passlib[bcrypt]
alembic
python-multipart
bcrypt==3.2.0
httpx
python-dateutil

# Phase 3: LangGraph Agent
langchain==0.3.25
langchain-groq==0.2.2
langchain-openai==0.1.24
langgraph==0.2.55
langgraph-checkpoint-sqlite==0.1.0
python-dotenv==1.0.1

# Phase 4: Anomaly Detection Engine
celery[redis]==5.5.3
redis==5.0.8
flower==2.0.1
kombu==5.4.12
pandas==2.2.0
numpy==1.26.4
scipy==1.12.0
